refactor: extract routes
